+++ apiphany-core/src/main/java/org/apiphany/lang/accumulator/CompositeAccumulator.java
@@ -15,6 +15,9 @@ import org.morphix.lang.Nullables;
  */
 public class CompositeAccumulator extends Accumulator<Object> {
 
+	/**
+	 * List of accumulators contained in this composite.
+	 */
 	private final List<Accumulator<?>> accumulators = new ArrayList<>();
 
 	/**